Changes to main block 1 Implementing solution for calculateArithmeticMean Implementing C1 to F1 ( C = 0) Implementing C2 to F2 ( C = 100) Implementing C3 to F3 ( C = 0) Implementing division program Implementing HoursToMinutesAndSecondsConvertor Implementing PercentCalculator Implementing SalaryCalculator test commit Changing in ArithmeticMeanCalculator.java "double a" Rename HoursToMinutesAndSecondsConvertor.java class

+++ src/main/java/part1/ArithmeticMeanCalculator.java
@@ -1,7 +1,7 @@
 package part1;
 
 public class ArithmeticMeanCalculator {
-    public static double calculateArithmeticMean(int a, double b, double c) {
+    public static double calculateArithmeticMean(double a, double b, double c) {
         // Calculate the mean
         return (a + b + c) / 3;
     }
